New function for loading FMOD configurations and ability for not starting FMOD on game start
As feedback from Ced, moved the code for loading the config file into its own function so new configurations can be called during runtime. Also created the option to toggle initialising FMOD on game start or to initialise it manually at a custom time. Calling a function to interact with FMOD before it's initialised will result in a crash.